requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
requests==2.32.3
gunicorn==23.0.0
Pillow==10.4.0